% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{bc_boxes}
\alias{bc_boxes}
\title{Extract box positions}
\usage{
bc_boxes(x, offset = 0L, fcol = "filterString")
}
\arguments{
\item{x}{An \code{MSnExp} object}

\item{offset}{\code{numeric(1)} defining the offset to remove/add to
the start/end of the box. Default is \code{0}, i.e. to leave them
as is.}

\item{fcol}{The name of the feature variable containing the box
data. Default is \code{filterString} and is expected to be
\code{"FTMS + p NSI SIM msx ms [start_1-end_1, start_2-end_2, ..., start_n-end_n]"}
where \code{start} and \code{end} are box start and end positions.}
}
\value{
A list of data frames of length \code{length(x)} with start and
end box values.
}
\description{
This function extracts the box start and end positions for all
spectra in an \code{MSnExp} object. The first box coordinates are not
returned, as they corresond to either the full spectrum (for
non-boxcar MS1 spectra) or the first start and last end box
positions for a boxcar MS1 spectrum. See the \emph{MSnbaseBoxCar}
vignette for example code.
}
