% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{bc_groups}
\alias{bc_groups}
\title{Define boxcar groups}
\usage{
bc_groups(x, fcol = "filterString")
}
\arguments{
\item{x}{An \code{MSnExp} object containing a set of boxcar spectra.}

\item{fcol}{The name of the feature variable containing the box
data. Default is \code{filterString}. See the \code{bc_boxes} function
for details.}
}
\value{
A \code{numeric} of length \code{length(x)} with the boxcar groups.
}
\description{
This function takes an \code{MSnExp} as input and defines the boxcar
groups, i.e. the set of boxcar spectra that match the same full
MS1 spectrum. Grouping for full (non-boxcar) spectra are set the
\code{NA}. Note that the presence of full (non-boxcar) spectra is
required to define the groups, and it is assumed that boxcar
spectra following a full MS1 spectrum below to the same group. The
output of this function is best used to create a new feature
variable column used when merging boxcar spectra.
}
\details{
TODO: make this an endomorphic function that return an MSnExp
object
}
\author{
Laurent Gatto
}
