% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{bc_is_boxcar}
\alias{bc_is_boxcar}
\title{Identifies boxcar spectra}
\usage{
bc_is_boxcar(x, fcol = "filterString")
}
\arguments{
\item{x}{An \code{MSnExp} object containing a set of boxcar spectra.}

\item{fcol}{The name of the feature variable containing the box
data. Default is \code{filterString}. See the \code{bc_boxes} function
for details.}
}
\value{
A \code{logical} of length \code{length(x)} indicating of a spectrum
is a boxcar spectrum.
}
\description{
This function uses the identification of boxcar boxes by the
\code{bc_boxes} function to determine if the spectrum is indeed a
boxcar spetrum. See the \emph{MSnbaseBoxCar} vignette for example code.
}
\author{
Laurent Gatto
}
