% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bc_plot}
\alias{bc_plot}
\alias{bc_plotly}
\title{Overlay boxcar spectra}
\usage{
bc_plot(x)

bc_plotly(x)
}
\arguments{
\item{x}{An \code{MSnExp} object containing a set of boxcar spectra.}
}
\value{
Used for its side effects. Invisibly returns a \code{ggplot} or
\code{plotly} object.
}
\description{
Takes an experiment with a set of boxcar spectra and plots them as
a single spectrum, colouring the peaks based on their origin.
}
\author{
Laurent Gatto
}
