% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{bc_zero_out_box}
\alias{bc_zero_out_box}
\title{Set peaks outside of boxes to zero}
\usage{
bc_zero_out_box(x, offset = 0L)
}
\arguments{
\item{x}{x An \code{MSnExp} object containing a set of boxcar spectra.}

\item{offset}{\code{numeric(1)} defining the offset to remove/add to
the start/end of the box. Default is \code{0}, i.e. to leave them
as is.}
}
\value{
An object of class \code{Spectra} with processed spectra.
}
\description{
Takes an MS experiment with boxcar spectra and sets any peak
outside of the boxes (+/- an optional offset) to zero. If the
spectrum is a fill, non-boxcar spectrum, it is left as is. See the
\emph{MSnbaseBoxCar} vignette for example code.
}
\author{
Laurent Gatto
}
