% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{boxcarCombine}
\alias{boxcarCombine}
\title{Combine boxcar spectra}
\usage{
boxcarCombine(x)
}
\arguments{
\item{x}{A list of spectra}
}
\description{
A function to be passed to \code{combineSpectra} to combine boxcar
spectra. This function simply sums intensities and is provided as
an example. See the \emph{MSnbaseBoxCar} vignette for example code.
}
