% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcar.R
\name{filterBoxCar}
\alias{filterBoxCar}
\title{Filter BoxCar spectra}
\usage{
filterBoxCar(x, fcol = "bc_groups")
}
\arguments{
\item{x}{An instance of class \code{MSnExp}.}

\item{fcol}{\code{character(1)} indicating the feature variable
defining the BoxCar spectra, typically generated by
\code{bc_groups()}. Default is \code{bc_groups}.}
}
\value{
An \code{MSnExp} containing only BoxCar spectra.
}
\description{
This function retains only BoxCar spectra in an \code{MSnExp}
instance. BoxCar spectra are defined as those that don't have an
\code{NA} in the feature variable \code{bc_groups}.
}
\author{
Laurent Gatto
}
